+++ CHANGELOG.md
@@ -4,7 +4,9 @@ CHANGELOG](http://keepachangelog.com/) for how to update this file. This project
 adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- Add missing Backend::Server#check_in specs
+### Added
+- Added missing Backend::Server#check_in specs
+- Added flag `--skip-rails-load` to cli commands for optionally skipping Rails initialization when running from a Rails root.
 
 ## [4.0.0] - 2018-08-21
 ### Added